Add transaction assembly strategies for different transfer types
